Restructure backend
