+++ mongodb-backend/tsconfig.json
@@ -3,6 +3,8 @@
       "module": "commonjs",
       "target": "es6",
       "rootDir": "./",
-      "esModuleInterop": true
+      "outDir": "./build",
+      "esModuleInterop": true,
+      "strict": true,
     }
 }
